legal-auditor/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
llama-index==0.11.0
llama-index-llms-ollama==0.3.0
llama-index-embeddings-ollama==0.3.0
llama-index-vector-stores-chroma==0.2.0
chromadb==0.5.0
pypdf>=4.0.1, <5.0.0
python-docx==1.1.0
aiosqlite==0.20.0
pydantic==2.9.0
fpdf2==2.8.2
razorpay
stripe==10.11.0
passlib[bcrypt]==1.7.4
bcrypt
python-jose[cryptography]==3.3.0
pydantic[email]
python-dotenv